[fix] merge conflict

+++ src/main/java/com/kt/edu/thirdproject/repository/HashtagRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HashtagRepository extends JpaRepository<Hashtag, Integer>{
-
+public interface HashtagRepository extends JpaRepository<Hashtag, Long> {
 }
